Move AuthenticationSortBy to model package
